Revert "Working pointstamped to navsatfix converter. Need to display navigator on rviz"
This reverts commit b85c34686a8ccb31d92b46620c5dba24b2bc6d85.
