+++ rviz_satellite/launch/demo.rviz
@@ -7,7 +7,6 @@ Panels:
         - /Global Options1
         - /Status1
         - /AerialMapDisplay1
-        - /Marker1
       Splitter Ratio: 0.5
     Tree Height: 581
   - Class: rviz/Selection
@@ -45,14 +44,6 @@ Visualization Manager:
       Topic: /satellite/navsatfix
       Value: true
       Zoom: 18
-    - Class: rviz/Marker
-      Enabled: true
-      Marker Topic: satellite/navigator_marker
-      Name: Marker
-      Namespaces:
-        {}
-      Queue Size: 100
-      Value: true
   Enabled: true
   Global Options:
     Background Color: 48; 48; 48
@@ -78,7 +69,7 @@ Visualization Manager:
   Views:
     Current:
       Class: rviz/Orbit
-      Distance: 293.635345
+      Distance: 478.777771
       Enable Stereo Rendering:
         Stereo Eye Separation: 0.0599999987
         Stereo Focal Distance: 1
@@ -93,10 +84,10 @@ Visualization Manager:
       Invert Z Axis: false
       Name: Current View
       Near Clip Distance: 0.00999999978
-      Pitch: 0.774796844
+      Pitch: 1.56979632
       Target Frame: <Fixed Frame>
       Value: Orbit (rviz)
-      Yaw: 4.99360132
+      Yaw: 4.65859365
     Saved: ~
 Window Geometry:
   Displays:
