+++ rviz_satellite/src/navigator_marker.h
@@ -1,0 +1,3 @@
+#include <ros/ros.h>
+#include <visualization_msgs/Marker.h>
+#include <string>
