+++ using_markers/CMakeLists.txt
@@ -196,5 +196,5 @@ include_directories(
 
 ## Add folders to be run by python nosetests
 # catkin_add_nosetests(test)
-add_executable(basic_shapes src/basic_shapes.cpp)
-target_link_libraries(basic_shapes ${catkin_LIBRARIES})
+add_executable(navigator_marker src/navigator_marker.cpp)
+target_link_libraries(navigator_marker ${catkin_LIBRARIES})
